Split out mapping helpers into separate files
