+++ sdk/typescript/src/graphql/client.ts
@@ -2,11 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { TypedDocumentNode } from '@graphql-typed-document-node/core';
-import { fromB64 } from '@mysten/bcs';
 import { print } from 'graphql';
 import type { DocumentNode } from 'graphql';
 
-import { bcs } from '../bcs/index.js';
 import type { PaginationArguments, SuiClientOptions } from '../client/client.js';
 import { SuiClient } from '../client/client.js';
 import type {
@@ -30,12 +28,8 @@ import type {
 	CommitteeInfo,
 	DelegatedStake,
 	DevInspectResults,
-	DisplayFieldsResponse,
 	DryRunTransactionBlockResponse,
-	ExecutionStatus,
 	MoveStruct,
-	MoveValue,
-	ObjectOwner,
 	ObjectRead,
 	PaginatedCoins,
 	PaginatedEvents,
@@ -44,17 +38,13 @@ import type {
 	ProtocolConfig,
 	ProtocolConfigValue,
 	SuiEvent,
-	SuiMoveAbility,
 	SuiMoveFunctionArgType,
 	SuiMoveNormalizedFunction,
 	SuiMoveNormalizedModule,
 	SuiMoveNormalizedStruct,
-	SuiMoveNormalizedType,
 	SuiObjectResponse,
 	SuiSystemStateSummary,
 	SuiTransactionBlockResponse,
-	SuiTransactionBlockResponseOptions,
-	SuiValidatorSummary,
 	TransactionEffects,
 	ValidatorsApy,
 } from '../client/types/generated.js';
@@ -94,17 +84,10 @@ import type {
 	SubscribeTransactionParams,
 	TryGetPastObjectParams,
 } from '../client/types/params.js';
-import { normalizeStructTag, normalizeSuiAddress, parseStructTag } from '../utils/sui-types.js';
+import { normalizeStructTag, parseStructTag } from '../utils/sui-types.js';
 import type {
 	ObjectFilter,
 	QueryEventsQueryVariables,
-	Rpc_Move_Function_FieldsFragment,
-	Rpc_Move_Module_FieldsFragment,
-	Rpc_Move_Object_FieldsFragment,
-	Rpc_Move_Struct_FieldsFragment,
-	Rpc_Object_FieldsFragment,
-	Rpc_Transaction_FieldsFragment,
-	Rpc_Validator_FieldsFragment,
 	TransactionBlockKindInput,
 } from './generated.js';
 import {
@@ -144,6 +127,17 @@ import {
 	TryGetPastObjectDocument,
 	TypedDocumentString,
 } from './generated.js';
+import { formatDisplay } from './mappers/display.js';
+import {
+	mapNormalizedMoveFunction,
+	mapNormalizedMoveModule,
+	mapNormalizedMoveStruct,
+	moveDataToRpcContent,
+} from './mappers/move.js';
+import { mapGraphQLMoveObjectToRpcObject, mapGraphQLObjectToRpcObject } from './mappers/object.js';
+import { mapGraphQLTransactionBlockToRpcTransactionBlock } from './mappers/transaction-block.js';
+import { isNumericString, toShortTypeString } from './mappers/util.js';
+import { mapGraphQlValidatorToRpcValidator } from './mappers/validator.js';
 
 export type GraphQLDocument<
 	Result = Record<string, unknown>,
@@ -1297,601 +1291,4 @@ class GraphQLResponseError extends Error {
 		super(error.message);
 		this.locations = error.locations;
 	}
-}
-
-function mapGraphQLObjectToRpcObject(
-	object: Rpc_Object_FieldsFragment,
-	options: { showBcs?: boolean | null } = {},
-): NonNullable<SuiObjectResponse['data']> {
-	return {
-		bcs: options?.showBcs
-			? {
-					dataType: 'moveObject' as const,
-					bcsBytes: object.asMoveObject?.contents?.bcs,
-					hasPublicTransfer: object.asMoveObject?.hasPublicTransfer!,
-					version: object.version as unknown as string, // RPC type is wrong here
-					type: toShortTypeString(object.asMoveObject?.contents?.type.repr!),
-			  }
-			: undefined,
-		content: {
-			dataType: 'moveObject' as const,
-			fields: moveDataToRpcContent(
-				object.asMoveObject?.contents?.data!,
-				object.asMoveObject?.contents?.type.layout!,
-			) as MoveStruct,
-			hasPublicTransfer: object.asMoveObject?.hasPublicTransfer!,
-			type: toShortTypeString(object.asMoveObject?.contents?.type.repr!),
-		},
-		digest: object.digest!,
-		display: formatDisplay(object),
-		objectId: object.objectId,
-		owner: mapGraphQLOwnerToRpcOwner(object.owner),
-		previousTransaction: object.previousTransactionBlock?.digest,
-		storageRebate: object.storageRebate,
-		type: toShortTypeString(object.asMoveObject?.contents?.type.repr!),
-		version: String(object.version),
-	};
-}
-
-function mapGraphQLMoveObjectToRpcObject(
-	object: Rpc_Move_Object_FieldsFragment,
-	options: { showBcs?: boolean | null } = {},
-): NonNullable<SuiObjectResponse['data']> {
-	return {
-		bcs: options?.showBcs
-			? {
-					dataType: 'moveObject' as const,
-					bcsBytes: object?.contents?.bcs,
-					hasPublicTransfer: object?.hasPublicTransfer!,
-					version: object.version as unknown as string, // RPC type is wrong here
-					type: toShortTypeString(object?.contents?.type.repr!),
-			  }
-			: undefined,
-		content: {
-			dataType: 'moveObject' as const,
-			fields: moveDataToRpcContent(
-				object?.contents?.data!,
-				object?.contents?.type.layout!,
-			) as MoveStruct,
-			hasPublicTransfer: object?.hasPublicTransfer!,
-			type: toShortTypeString(object?.contents?.type.repr!),
-		},
-		digest: object.digest!,
-		display: formatDisplay(object),
-		objectId: object.objectId,
-		owner: mapGraphQLOwnerToRpcOwner(object.owner),
-		previousTransaction: object.previousTransactionBlock?.digest,
-		storageRebate: object.storageRebate,
-		type: toShortTypeString(object?.contents?.type.repr!),
-		version: String(object.version),
-	};
-}
-
-function mapGraphQLOwnerToRpcOwner(owner: Rpc_Object_FieldsFragment['owner']): ObjectOwner | null {
-	switch (owner?.__typename) {
-		case 'AddressOwner':
-			return owner.owner?.asObject
-				? {
-						ObjectOwner: owner.owner?.asObject.address!,
-				  }
-				: {
-						AddressOwner: owner.owner?.asAddress?.address!,
-				  };
-		case 'Parent':
-			return {
-				ObjectOwner: owner.parent?.address,
-			};
-		case 'Shared': {
-			return {
-				Shared: {
-					initial_shared_version: String(owner.initialSharedVersion),
-				},
-			};
-		}
-		case 'Immutable':
-			return 'Immutable';
-	}
-
-	return null;
-}
-
-function mapGraphQLTransactionBlockToRpcTransactionBlock(
-	transactionBlock: Rpc_Transaction_FieldsFragment,
-	options?: SuiTransactionBlockResponseOptions | null,
-): SuiTransactionBlockResponse {
-	const deletedChanges = transactionBlock.effects?.objectChanges?.nodes
-		?.filter((change) => change?.idDeleted === true)
-		.map((change) => ({
-			digest: change?.inputState?.digest!,
-			version: String(change?.inputState?.version),
-			objectId: change?.inputState?.address,
-		}));
-	const createdChanges = transactionBlock.effects?.objectChanges?.nodes
-		?.filter((change) => change?.idCreated === true)
-		.map((change) => ({
-			owner: mapGraphQLOwnerToRpcOwner(change?.outputState?.owner)!,
-			reference: {
-				digest: change?.outputState?.digest!,
-				version: change?.outputState?.version as unknown as string, // RPC type is wrong here
-				objectId: change?.outputState?.address,
-			},
-		}));
-
-	return {
-		balanceChanges: transactionBlock.effects?.balanceChanges?.nodes.map((balanceChange) => ({
-			amount: balanceChange?.amount,
-			coinType: toShortTypeString(balanceChange?.coinType?.repr),
-			owner: balanceChange.owner?.asObject?.address
-				? {
-						ObjectOwner: balanceChange.owner?.asObject?.address,
-				  }
-				: {
-						AddressOwner: balanceChange.owner?.asAddress?.address!,
-				  },
-		})),
-		checkpoint: transactionBlock.effects?.checkpoint?.sequenceNumber.toString(),
-		timestampMs: new Date(transactionBlock.effects?.timestamp).getTime().toString(),
-		digest: transactionBlock.digest!,
-		effects: options?.showEffects
-			? {
-					...(createdChanges?.length ? { created: createdChanges } : {}),
-					...(deletedChanges?.length ? { deleted: deletedChanges } : {}),
-					dependencies: transactionBlock.effects?.dependencies?.nodes.map((dep) => dep?.digest!),
-					executedEpoch: String(transactionBlock.effects?.executedEpoch?.epochId),
-					gasObject: {
-						owner: mapGraphQLOwnerToRpcOwner(
-							transactionBlock.effects?.gasEffects?.gasObject?.owner,
-						)!,
-						reference: {
-							digest: transactionBlock.effects?.gasEffects?.gasObject?.digest!,
-							version: transactionBlock.effects?.gasEffects?.gasObject
-								?.version as unknown as string, // RPC type is wrong here
-							objectId: transactionBlock.effects?.gasEffects?.gasObject?.address,
-						},
-					},
-					gasUsed: {
-						computationCost: transactionBlock.effects?.gasEffects?.gasSummary?.computationCost,
-						nonRefundableStorageFee:
-							transactionBlock.effects?.gasEffects?.gasSummary?.nonRefundableStorageFee,
-						storageCost: transactionBlock.effects?.gasEffects?.gasSummary?.storageCost,
-						storageRebate: transactionBlock.effects?.gasEffects?.gasSummary?.storageRebate,
-					},
-					messageVersion: 'v1' as const,
-					modifiedAtVersions: transactionBlock.effects?.objectChanges?.nodes
-						?.filter((change) => !change?.idCreated && !change?.idDeleted)
-						?.map((change) => ({
-							objectId: change?.inputState?.address,
-							sequenceNumber: String(change?.inputState?.version),
-						})),
-					mutated: transactionBlock.effects?.objectChanges?.nodes
-						?.filter((change) => !change?.idCreated && !change?.idDeleted)
-						?.map((change) => ({
-							owner: mapGraphQLOwnerToRpcOwner(change?.outputState?.owner)!,
-							reference: {
-								digest: change?.outputState?.digest!,
-								version: change?.outputState?.version as unknown as string,
-								objectId: change?.outputState?.address,
-							},
-						})),
-
-					status: { status: transactionBlock.effects?.status?.toLowerCase() } as ExecutionStatus,
-					transactionDigest: transactionBlock.digest!,
-					// sharedObjects: [], // TODO
-					// unwrapped: [], // TODO
-					// unwrappedThenDeleted: [], // TODO
-					// wrapped: [], // TODO
-			  }
-			: undefined,
-		// errors: [], // TODO
-		events: options?.showEvents
-			? transactionBlock.effects?.events?.nodes.map((event) => ({
-					bcs: event.bcs,
-					id: 'TODO' as never, // TODO: turn id into an object
-					packageId: event.sendingModule?.package.address!,
-					parsedJson: event.json ? JSON.parse(event.json) : undefined,
-					sender: event.sender?.address,
-					timestampMs: new Date(event.timestamp).getTime().toString(),
-					transactionModule: 'TODO',
-					type: toShortTypeString(event.type?.repr)!,
-			  })) ?? []
-			: undefined,
-		rawTransaction: options?.showRawInput ? transactionBlock.rawTransaction : undefined,
-		transaction: options?.showInput &&
-			transactionBlock.rawTransaction && {
-				data: bcs.SenderSignedData.parse(fromB64(transactionBlock.rawTransaction))[0].intentMessage
-					.value.V1,
-			},
-		objectChanges: options?.showObjectChanges
-			? transactionBlock.effects?.objectChanges?.nodes
-					?.map((change) =>
-						change?.idDeleted
-							? {
-									digest: change?.inputState?.digest!,
-									objectId: change?.inputState?.address,
-									owner: mapGraphQLOwnerToRpcOwner(change.inputState?.owner),
-									objectType: toShortTypeString(
-										change?.inputState?.asMoveObject?.contents?.type.repr,
-									),
-									sender: transactionBlock.sender?.address!,
-									type: 'deleted' as const,
-									version: change?.inputState?.version.toString()!,
-							  }
-							: {
-									digest: change?.outputState?.digest!,
-									objectId: change?.outputState?.address,
-									owner: mapGraphQLOwnerToRpcOwner(change.outputState?.owner)!,
-									objectType: toShortTypeString(
-										change?.outputState?.asMoveObject?.contents?.type.repr,
-									),
-									previousVersion: change?.inputState?.version.toString()!,
-									sender: transactionBlock.sender?.address,
-									type: change?.idCreated ? ('created' as const) : ('mutated' as const),
-									version: change?.outputState?.version.toString()!,
-							  },
-					)
-					.sort((a, b) => {
-						if (a.type === 'created' && b.type === 'deleted') {
-							return -1;
-						} else if (a.type === 'deleted' && b.type === 'created') {
-							return 1;
-						}
-						return 0;
-					})
-			: undefined,
-	};
-}
-
-type MoveData =
-	| { Address: number[] }
-	| { UID: number[] }
-	| { ID: number[] }
-	| { Bool: boolean }
-	| { Number: string }
-	| { String: string }
-	| { Vector: MoveData[] }
-	| { Option: MoveData | null }
-	| { Struct: [{ name: string; value: MoveData }] };
-
-type MoveTypeLayout =
-	| 'address'
-	| 'bool'
-	| 'u8'
-	| 'u16'
-	| 'u32'
-	| 'u64'
-	| 'u128'
-	| 'u256'
-	| { vector: MoveTypeLayout }
-	| {
-			struct: {
-				type: string;
-				fields: { name: string; layout: MoveTypeLayout }[];
-			};
-	  };
-
-function moveDataToRpcContent(data: MoveData, layout: MoveTypeLayout): MoveValue {
-	if ('Address' in data) {
-		return normalizeSuiAddress(
-			data.Address.map((byte) => byte.toString(16).padStart(2, '0')).join(''),
-		);
-	}
-
-	if ('UID' in data) {
-		return {
-			id: normalizeSuiAddress(data.UID.map((byte) => byte.toString(16).padStart(2, '0')).join('')),
-		};
-	}
-
-	if ('ID' in data) {
-		return normalizeSuiAddress(data.ID.map((byte) => byte.toString(16).padStart(2, '0')).join(''));
-	}
-
-	if ('Bool' in data) {
-		return data.Bool;
-	}
-
-	if ('Number' in data) {
-		return layout === 'u64' || layout === 'u128' || layout === 'u256'
-			? String(data.Number)
-			: Number.parseInt(data.Number, 10);
-	}
-
-	if ('String' in data) {
-		return data.String;
-	}
-
-	if ('Vector' in data) {
-		if (typeof layout !== 'object' || !('vector' in layout)) {
-			throw new Error(`Invalid layout for data: ${JSON.stringify(data)}}`);
-		}
-		const itemLayout = layout.vector;
-		return data.Vector.map((item) => moveDataToRpcContent(item, itemLayout));
-	}
-
-	if ('Option' in data) {
-		return data.Option && moveDataToRpcContent(data.Option, layout);
-	}
-
-	if ('Struct' in data) {
-		const result: MoveStruct = {};
-
-		if (typeof layout !== 'object' || !('struct' in layout)) {
-			throw new Error(`Invalid layout for data: ${JSON.stringify(data)}}`);
-		}
-
-		data.Struct.forEach((item, index) => {
-			const { name, layout: itemLayout } = layout.struct.fields[index];
-
-			result[name] = moveDataToRpcContent(item.value, itemLayout);
-		});
-
-		// https://github.com/MystenLabs/sui/blob/5849f6845a3ab9fdb4c17523994adad461478a4c/crates/sui-json-rpc-types/src/sui_move.rs#L481
-		const tag = parseStructTag(layout.struct.type);
-		const structName = `${toShortTypeString(tag.address)}::${tag.module}::${tag.name}`;
-
-		switch (structName) {
-			case '0x1::string::String':
-			case '0x1::ascii::String':
-				return result['bytes'];
-			case '0x2::url::Url':
-				return result['url'];
-			case '0x2::object::ID':
-				return result['bytes'];
-			case '0x2::object::UID':
-				return {
-					id: result['id'] as string,
-				};
-			case '0x2::balance::Balance':
-				return result['value'];
-			case '0x1::option::Option':
-				return (result['vec'] as MoveValue[])[0] ?? null;
-		}
-
-		return result;
-	}
-
-	throw new Error('Invalid move data: ' + JSON.stringify(data));
-}
-
-function mapGraphQlValidatorToRpcValidator(
-	validator: Rpc_Validator_FieldsFragment,
-): SuiValidatorSummary {
-	return {
-		commissionRate: validator.commissionRate?.toString()!,
-		description: validator.description!,
-		exchangeRatesId: validator.exchangeRates?.address!,
-		exchangeRatesSize: 'TODO',
-		gasPrice: validator.gasPrice,
-		imageUrl: validator.imageUrl!,
-		name: validator.name!,
-		netAddress: validator.credentials?.netAddress!,
-		networkPubkeyBytes: validator.credentials?.networkPubKey!,
-		nextEpochCommissionRate: validator.nextEpochCommissionRate?.toString()!,
-		nextEpochGasPrice: validator.nextEpochGasPrice,
-		nextEpochNetAddress: validator.nextEpochCredentials?.netAddress,
-		nextEpochNetworkPubkeyBytes: validator.nextEpochCredentials?.networkPubKey,
-		nextEpochP2pAddress: validator.nextEpochCredentials?.p2PAddress,
-		nextEpochPrimaryAddress: validator.nextEpochCredentials?.primaryAddress,
-		nextEpochProofOfPossession: validator.nextEpochCredentials?.proofOfPossession,
-		nextEpochProtocolPubkeyBytes: validator.nextEpochCredentials?.protocolPubKey,
-		nextEpochStake: validator.nextEpochStake!,
-		nextEpochWorkerAddress: validator.nextEpochCredentials?.workerAddress,
-		nextEpochWorkerPubkeyBytes: validator.nextEpochCredentials?.workerPubKey,
-		operationCapId: validator.operationCap?.address!,
-		p2pAddress: validator.credentials?.p2PAddress!,
-		pendingTotalSuiWithdraw: validator.pendingTotalSuiWithdraw,
-		pendingPoolTokenWithdraw: validator.pendingPoolTokenWithdraw,
-		poolTokenBalance: validator.poolTokenBalance,
-		pendingStake: validator.pendingStake,
-		primaryAddress: validator.credentials?.primaryAddress!,
-		projectUrl: validator.projectUrl!,
-		proofOfPossessionBytes: validator.credentials?.proofOfPossession,
-		protocolPubkeyBytes: validator.credentials?.protocolPubKey,
-		rewardsPool: validator.rewardsPool,
-		stakingPoolId: validator.stakingPool?.address!,
-		stakingPoolActivationEpoch: validator.stakingPoolActivationEpoch?.toString(),
-		stakingPoolSuiBalance: validator.stakingPoolSuiBalance,
-		suiAddress: validator.address.address,
-		votingPower: validator.votingPower?.toString()!,
-		workerAddress: validator.credentials?.workerAddress!,
-		workerPubkeyBytes: validator.credentials?.workerPubKey,
-	};
-}
-
-function toShortTypeString<T extends string | null | undefined>(type?: T): T {
-	return type?.replace(/0x0+/g, '0x') as T;
-}
-
-function isNumericString(value: string) {
-	return /^-?\d+$/.test(value);
-}
-
-type OpenMoveTypeSignatureBody =
-	| 'address'
-	| 'bool'
-	| 'u8'
-	| 'u16'
-	| 'u32'
-	| 'u64'
-	| 'u128'
-	| 'u256'
-	| { vector: OpenMoveTypeSignatureBody }
-	| {
-			datatype: {
-				package: string;
-				module: string;
-				type: string;
-				typeParameters?: [OpenMoveTypeSignatureBody];
-			};
-	  }
-	| { typeParameter: number };
-
-function mapOpenMoveType(type: { ref?: '&' | '&mut'; body: OpenMoveTypeSignatureBody }) {
-	const body = mapNormalizedType(type.body);
-
-	if (type.ref === '&') {
-		return {
-			Reference: body,
-		};
-	}
-
-	if (type.ref === '&mut') {
-		return {
-			MutableReference: body,
-		};
-	}
-
-	return body;
-}
-
-function mapNormalizedType(type: OpenMoveTypeSignatureBody): SuiMoveNormalizedType {
-	switch (type) {
-		case 'address':
-			return 'Address';
-		case 'bool':
-			return 'Bool';
-		case 'u8':
-			return 'U8';
-		case 'u16':
-			return 'U16';
-		case 'u32':
-			return 'U32';
-		case 'u64':
-			return 'U64';
-		case 'u128':
-			return 'U128';
-		case 'u256':
-			return 'U256';
-	}
-
-	if ('vector' in type) {
-		return {
-			Vector: mapNormalizedType(type.vector),
-		};
-	}
-
-	if ('typeParameter' in type) {
-		return {
-			TypeParameter: type.typeParameter,
-		};
-	}
-
-	if ('datatype' in type) {
-		return {
-			Struct: {
-				address: toShortTypeString(type.datatype.package),
-				module: type.datatype.module,
-				name: type.datatype.type,
-				typeArguments: type.datatype.typeParameters?.map(mapNormalizedType) ?? [],
-			},
-		};
-	}
-
-	throw new Error('Invalid type');
-}
-
-function mapNormalizedMoveFunction(
-	fn: Rpc_Move_Function_FieldsFragment,
-): SuiMoveNormalizedFunction {
-	return {
-		visibility: `${fn.visibility?.[0]}${fn.visibility?.slice(1).toLowerCase()}` as never,
-		isEntry: fn.isEntry!,
-		typeParameters:
-			fn.typeParameters?.map((param) => ({
-				abilities:
-					param.constraints?.map(
-						(constraint) =>
-							`${constraint[0]}${constraint.slice(1).toLowerCase()}` as SuiMoveAbility,
-					) ?? [],
-			})) ?? [],
-		return: fn.return?.map((param) => mapOpenMoveType(param.signature)) ?? [],
-		parameters: fn.parameters?.map((param) => mapOpenMoveType(param.signature)) ?? [],
-	};
-}
-
-function mapNormalizedMoveStruct(struct: Rpc_Move_Struct_FieldsFragment): SuiMoveNormalizedStruct {
-	return {
-		abilities: {
-			abilities:
-				struct.abilities?.map(
-					(ability) => `${ability[0]}${ability.slice(1).toLowerCase()}` as SuiMoveAbility,
-				) ?? [],
-		},
-		fields:
-			struct.fields?.map((field) => ({
-				name: field.name,
-				type: mapOpenMoveType(field.type?.signature),
-			})) ?? [],
-		typeParameters:
-			struct.typeParameters?.map((param) => ({
-				isPhantom: param.isPhantom!,
-				constraints: {
-					abilities: param.constraints?.map(
-						(constraint) =>
-							`${constraint[0]}${constraint.slice(1).toLowerCase()}` as SuiMoveAbility,
-					),
-				},
-			})) ?? [],
-	};
-}
-
-function mapNormalizedMoveModule(
-	module: Rpc_Move_Module_FieldsFragment,
-	address: string,
-): SuiMoveNormalizedModule {
-	const exposedFunctions: Record<string, SuiMoveNormalizedFunction> = {};
-	const structs: Record<string, SuiMoveNormalizedStruct> = {};
-
-	module.functions?.nodes
-		.filter((func) => func.visibility === 'PUBLIC' || func.isEntry || func.visibility === 'FRIEND')
-		.forEach((func) => {
-			exposedFunctions[func.name] = mapNormalizedMoveFunction(func);
-		});
-
-	module.structs?.nodes.forEach((struct) => {
-		structs[struct.name] = mapNormalizedMoveStruct(struct);
-	});
-
-	return {
-		address: toShortTypeString(address),
-		name: module.name,
-		fileFormatVersion: module.fileFormatVersion,
-		friends:
-			module.friends.nodes?.map((friend) => ({
-				address: toShortTypeString(friend.package.address),
-				name: friend.name,
-			})) ?? [],
-		structs,
-		exposedFunctions,
-	};
-}
-
-function formatDisplay(object: {
-	display?:
-		| {
-				key: string;
-				value?: string | null | undefined;
-				error?: string | null | undefined;
-		  }[]
-		| null;
-}) {
-	let display: DisplayFieldsResponse = {
-		data: null,
-		error: null,
-	};
-
-	if (object.display) {
-		object.display.forEach((displayItem) => {
-			if (displayItem.error) {
-				display!.error = displayItem.error as never;
-			} else if (displayItem.value != null) {
-				if (!display!.data) {
-					display!.data = {};
-				}
-				display!.data[displayItem.key] = displayItem.value;
-			}
-		});
-	}
-
-	return display;
 }
