+++ sdk/typescript/src/graphql/mappers/util.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+export function toShortTypeString<T extends string | null | undefined>(type?: T): T {
+	return type?.replace(/0x0+/g, '0x') as T;
+}
+
+export function isNumericString(value: string) {
+	return /^-?\d+$/.test(value);
+}
